+++ Items/Armour/Centurion/RavennaBoots.cs
@@ -27,9 +27,9 @@ namespace ArcaneOdyssey.Items.Armour.Centurion
 			}
 		}
 
-		public override int AODefense => 56;
-		public override int Size => AODefense / 20;
-		public override int AOAttkSpd => AODefense / 20;
+		public override ushort AODefense => 56;
+		public override short Size => (short)(AODefense / 20);
+		public override short AOAttkSpd => (short)(AODefense / 20);
 		public override ItemRarities Rarity => ItemRarities.Common;
 		public override int Value => 30;
