+++ Items/Armour/Centurion/RavennaHelm.cs
@@ -8,9 +8,9 @@ namespace ArcaneOdyssey.Items.Armour.Centurion
 	public class RavennaHelm : Base.Armour
 	{
 		public override ItemTiers ArmourTier => ItemTiers.Average;
-		public override int AODefense => 144;
-		public override int Size => AODefense / 17;
-		public override int AOAttkSpd => AODefense / 17;
+		public override ushort AODefense => 144;
+		public override short Size => (short)(AODefense / 17);
+		public override short AOAttkSpd => (short)(AODefense / 17);
 		public override ItemRarities Rarity => ItemRarities.Uncommon;
 		public override int Value => 55;
