+++ Items/Armour/RavennaNoble/EliusBoots.cs
@@ -9,8 +9,8 @@ namespace ArcaneOdyssey.Items.Armour.RavennaNoble
 		public override ItemTiers ArmourTier => ItemTiers.Average;
 		public override ItemRarities Rarity => ItemRarities.Uncommon;
 		public override int Value => 60;
-		public override int AOAgility => 11;
-		public override int AOPower => 9;
+		public override short AOAgility => 11;
+		public override short AOPower => 9;
 
 		public override void SetDefaults()
 		{
