+++ Items/Armour/RavennaNoble/EliusHelm.cs
@@ -22,7 +22,7 @@ namespace ArcaneOdyssey.Items.Armour.RavennaNoble
 		public override ItemRarities Rarity => ItemRarities.Uncommon;
 		public override int Value => 80;
 
-		public override int AOAgility => 9;
-		public override int AOPower => 7;
+		public override short AOAgility => 9;
+		public override short AOPower => 7;
 	}
 }
